"Because" overlay test buttons resize
This update has the "Because" overlay window working to the point were the size will respond appropriately to clicks so that they are all sized in proportion to total number of clicks. Also the "cancel" button is the only button that turns the overlay off. And the "redo" icon works to reset the buttons and reset the global totalClicks and data array.