Update CDDb_working - labeled buttons
The buttons are now fully functional, though there is something up with the percentage sign. It should be a ratio of the font-size of the numbers that is dynamically set in proportion to the click ratio but it seems to hit a floor in size it doesn't hit with the display_associated php it was modeled after. Might need to mess with spacing between number and percent sign as well. Small nits.